Cleaned up reducers and set vendor and feature types with FETCH_USER
