+++ src/redux/sagas/_root.saga.js
@@ -3,11 +3,11 @@ import loginSaga from './login.saga';
 import registrationSaga from './registration.saga';
 import userSaga from './user.saga';
 import userDetailsSaga from './user.details.saga';
-import chatSaga from './chat.saga';
+import messageSaga from './message.saga';
 import searchSaga from './search.saga';
 import dropdownSaga from './dropdown.saga';
 import vendorSaga from './vendor.saga';
-import filterSaga from './filter.saga';
+
 
 // rootSaga is the primary saga.
 // It bundles up all of the other sagas so our project can use them.
@@ -22,10 +22,9 @@ export default function* rootSaga() {
     registrationSaga(),
     userSaga(),
     userDetailsSaga(),
-    chatSaga(),
+    messageSaga(),
     searchSaga(),
     dropdownSaga(),
     vendorSaga(),
-    filterSaga(),
   ]);
 }
