+++ src/redux/sagas/login.saga.js
@@ -21,6 +21,7 @@ function* loginUser(action) {
     // after the user has logged in
     // get the user information from the server
     yield put({ type: 'FETCH_USER' });
+
   } catch (error) {
     console.log('Error with user login:', error);
     if (error.response.status === 401) {
