+++ src/redux/sagas/search.saga.js
@@ -1,24 +1,27 @@
 import axios from 'axios';
-import { put, takeLatest } from 'redux-saga/effects';
+import { put, takeLatest, takeEvery } from 'redux-saga/effects';
 
-function* runSearch(action) {
-  const searchTerm = action.payload.searchInput;
+// Creates categorySAGA Generator function
+function* searchSaga() {
+  yield takeEvery('FETCH_MATCHING_VENDORS', fetchMatchingVendors);
+}
 
-  console.log('runSearch searchTerm:', searchTerm);
+function* fetchMatchingVendors(action) {
+  const vendorTypeSelection = action.payload.typeId;
+  const specialFeatureSelection = action.payload.featureId;
+  const searchTerm = action.payload.searchTermInput;
+  console.log('searchTerm:', searchTerm);
+
+  console.log('start of fetchMatchingVendors', vendorTypeSelection, specialFeatureSelection, searchTerm);
 
   try {
-    // const searchResults = yield axios.get('/api/search', searchTerm);
-    const searchResults = yield axios.get(`/api/search/${searchTerm}`);
+    const response = yield axios.get(`/api/filter/typeId=${vendorTypeSelection}/featureId=${specialFeatureSelection}/searchTerm=${searchTerm}`);
 
-    yield put({ type: 'SET_SEARCH_RESULTS', payload: searchResults.data });
-  } catch (err) {
-    console.error('CLIENT - SAGAS - an error occurred fetching search results');
+    // yield put({ type: 'SET_MATCHING_VENDORS', payload: response.data });
+    yield put({ type: 'SET_SEARCH_RESULTS', payload: response.data });
+
+  } catch (error) {
+    console.log('fetchMatchingVendors get request failed', error);
   }
 }
-
-
-function* searchSaga() {
-  yield takeLatest('RUN_SEARCH', runSearch);
-}
-
 export default searchSaga;
