+++ src/redux/sagas/user.saga.js
@@ -19,6 +19,8 @@ function* fetchUser() {
     // with an id and username set the client-side user object to let
     // the client-side code know the user is logged in
     yield put({ type: 'SET_USER', payload: response.data });
+    yield put({ type: 'FETCH_VENDOR_TYPES' });
+    yield put({ type: 'FETCH_SPECIAL_FEATURES' });
   } catch (error) {
     console.log('User get request failed', error);
   }
